tkinter-demo/pyproject.toml
[project]
name = "tkinter-demo"
version = "0.1.0"
description = ""
authors = [
    {name = "GagMirz",email = "gmirzoyan12@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyserial (>=3.5,<4.0)"
]

[tool.poetry]
packages = [{include = "tkinter_demo", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
tx = "tkinter_demo.transmitter:main"
rx = "tkinter_demo.receiver:main"
